requirements.txt

fastapi
uvicorn
pydantic
pydantic-settings
pillow
python-multipart
requests
ultralytics
